Resource based authorization

+++ HotelsApi/Hotelss.Application/Hotels/Commands/DeleteHotel/DeleteHotelCommandHandler.cs
@@ -1,5 +1,7 @@
-﻿using Hotelss.Domain.Entities;
+﻿using Hotelss.Domain.Constants;
+using Hotelss.Domain.Entities;
 using Hotelss.Domain.Exceptions;
+using Hotelss.Domain.Interfaces;
 using Hotelss.Domain.Repositories;
 using MediatR;
 using Microsoft.Extensions.Logging;
@@ -7,7 +9,8 @@ using Microsoft.Extensions.Logging;
 namespace Hotelss.Application.Hotels.Commands.DeleteHotel;
 
 public class DeleteHotelCommandHandler(ILogger<DeleteHotelCommandHandler> logger,
-    IHotelsRepository hotelsRepository) : IRequestHandler<DeleteHotelCommand>
+    IHotelsRepository hotelsRepository,
+    IHotelAuthorizationService hotelAuthorizationService) : IRequestHandler<DeleteHotelCommand>
 {
     public async Task Handle(DeleteHotelCommand request, CancellationToken cancellationToken)
     {
@@ -16,6 +19,9 @@ public class DeleteHotelCommandHandler(ILogger<DeleteHotelCommandHandler> logger
 
         if ( hotel == null )
             throw new NotFoundException(nameof(Hotel), request.Id.ToString());
+
+        if (!hotelAuthorizationService.Authorize(hotel, ResourceOperation.Delete))
+            throw new ForbidException();
 
         await hotelsRepository.Delete(hotel);
     }
